change order of union types

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -4,6 +4,6 @@
 from typing import List, Union
 
 
-def sum_mixed_list(mxd_lst: List[Union[float, int]]) -> float:
+def sum_mixed_list(mxd_lst: List[Union[int, float]]) -> float:
     '''Return the sum of the mixed list.'''
     return sum(mxd_lst)
